+++ Tum4ik.JustClipboardManager/Constants/RegionNames.cs
@@ -2,6 +2,7 @@ namespace Tum4ik.JustClipboardManager.Constants;
 
 internal static class RegionNames
 {
-  public const string MainDialogContent = "MainDialogContent";
-  public const string MainDialogSettingsViewContent = "MainDialogSettingsViewContent";
+  public const string MainDialogContent = nameof(MainDialogContent);
+  public const string MainDialogSettingsViewContent = nameof(MainDialogSettingsViewContent);
+  public const string MainDialogPluginsViewContent = nameof(MainDialogPluginsViewContent);
 }