+++ Tum4ik.JustClipboardManager/Data/Dto/ClipDto.cs
@@ -1,0 +1,8 @@
+namespace Tum4ik.JustClipboardManager.Data.Dto;
+internal class ClipDto
+{
+  public required int Id { get; set; }
+  public required string PluginId { get; set; }
+  public required object RepresentationData { get; set; }
+  public string? SearchLabel { get; set; }
+}