+++ Tum4ik.JustClipboardManager/Data/Dto/SearchPluginInfoDto.cs
@@ -1,0 +1,7 @@
+namespace Tum4ik.JustClipboardManager.Data.Dto;
+
+internal class SearchPluginInfoDto : PluginInfoDto
+{
+  public required Uri DownloadLink { get; init; }
+  public bool IsInstalled { get; set; }
+}