+++ Tum4ik.JustClipboardManager/Views/Main/Plugins/PluginsInstalledView.xaml.cs
@@ -1,0 +1,11 @@
+namespace Tum4ik.JustClipboardManager.Views.Main.Plugins;
+/// <summary>
+/// Interaction logic for PluginsInstalledView.xaml
+/// </summary>
+public partial class PluginsInstalledView
+{
+  public PluginsInstalledView()
+  {
+    InitializeComponent();
+  }
+}