+++ Tum4ik.JustClipboardManager/Views/Main/Plugins/PluginsSearchView.xaml.cs
@@ -1,0 +1,11 @@
+namespace Tum4ik.JustClipboardManager.Views.Main.Plugins;
+/// <summary>
+/// Interaction logic for PluginsSearchView.xaml
+/// </summary>
+public partial class PluginsSearchView
+{
+  public PluginsSearchView()
+  {
+    InitializeComponent();
+  }
+}