+++ Tum4ik.JustClipboardManager/Views/Main/Plugins/PluginsView.xaml.cs
@@ -1,0 +1,11 @@
+namespace Tum4ik.JustClipboardManager.Views.Main.Plugins;
+/// <summary>
+/// Interaction logic for PluginsView.xaml
+/// </summary>
+public partial class PluginsView
+{
+  public PluginsView()
+  {
+    InitializeComponent();
+  }
+}